requirements.txt
SQLAlchemy~=2.0.21
python-dotenv~=1.0.0
fastapi~=0.103.2
psycopg2-binary~=2.9.9
uvicorn~=0.23.2
PyJWT~=2.8.0